Fix the logging of iframes
For whatever reason, beforeload doesn't work on children iframes
(they're new document roots or something). Since this is a really common
way of doing script injecting, we need a workaround.

We instead listen for elements getting added to the DOM. We catch all
iframes added in this way and scrape their src attribute.

However, some people (notable offenders include Facebook) insert their
iframes before setting a src. Additionally, we need to make sure that
the src attribute doesn't change behind our back. To do this, we inject
a script into the page's context, which modifies the .src property and
the setAttribute function on all iframes to notify us of all src
changes. This gets sent back to our extension for processing via various
hacks and trickery.